requirements.txt
fastapi
uvicorn
gunicorn
python-docx
python-multipart
groq
python-dotenv